Moved different widgets to 'home_widgets' folder
