+++ Flutter/flutter_catalogue/lib/pages/home_page.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_catalogue/models/catalogue.dart';
-// ignore: unused_import
-import 'package:flutter_catalogue/widgets/drawer.dart';
-// ignore: unused_import
-import 'package:flutter_catalogue/widgets/item_widget.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_catalogue/pages/home_widgets/catalogue_header.dart';
+import 'package:flutter_catalogue/pages/home_widgets/catalogue_list.dart';
 import 'package:flutter_catalogue/widgets/themes.dart';
 import 'dart:convert';
 import 'package:velocity_x/velocity_x.dart';
@@ -51,101 +49,11 @@ class _HomePageState extends State<HomePage> {
                   CatalogueModel.items.isNotEmpty)
                 CatalogueList().expand()
               else
-                Center(
-                  child: CircularProgressIndicator(),
-                )
+                CircularProgressIndicator().centered().expand(),
             ],
           ),
         ),
       ),
     );
-  }
-}
-
-class CatalogueHeader extends StatelessWidget {
-  const CatalogueHeader({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        "Catalogue App".text.xl4.bold.color(MyTheme.darkBluishColor).make(),
-        "Trending Products".text.xl2.make(),
-      ],
-    );
-  }
-}
-
-class CatalogueList extends StatelessWidget {
-  const CatalogueList({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return ListView.builder(
-      shrinkWrap: true,
-      itemCount: CatalogueModel.items.length,
-      itemBuilder: (context, index) {
-        final catalogue = CatalogueModel.items[index];
-        return CatalogueItem(catalogue: catalogue);
-      },
-    );
-  }
-}
-
-class CatalogueItem extends StatelessWidget {
-  final Item catalogue;
-
-  const CatalogueItem({Key? key, required this.catalogue}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return VxBox(
-      child: Row(
-        children: [
-          CatalogueImage(
-            image: catalogue.image,
-          ),
-          Expanded(
-              child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              catalogue.name.text.lg.color(MyTheme.darkBluishColor).bold.make(),
-              catalogue.desc.text /*textStyle(context.captionStyle)*/ .make(),
-              10.heightBox,
-              ButtonBar(
-                alignment: MainAxisAlignment.spaceBetween,
-                buttonPadding: EdgeInsets.zero,
-                children: [
-                  "\$${catalogue.price}".text.bold.xl.make(),
-                  ElevatedButton(
-                    onPressed: () {},
-                    style: ButtonStyle(
-                        backgroundColor:
-                            MaterialStateProperty.all(MyTheme.darkBluishColor),
-                        shape: MaterialStateProperty.all(StadiumBorder())),
-                    child: "Buy".text.make(),
-                  )
-                ],
-              ).pOnly(right: 8.0)
-            ],
-          ))
-        ],
-      ),
-    ).white.roundedLg.square(150).make().py16();
-  }
-}
-
-class CatalogueImage extends StatelessWidget {
-  final String image;
-
-  const CatalogueImage({Key? key, required this.image}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Image.network(
-      image,
-    ).box.rounded.p8.color(MyTheme.creamColor).make().p16().w40(context);
   }
 }
