+++ Flutter/flutter_catalogue/lib/pages/home_widgets/catalogue_header.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+import 'package:velocity_x/velocity_x.dart';
+import 'package:flutter_catalogue/widgets/themes.dart';
+
+class CatalogueHeader extends StatelessWidget {
+  const CatalogueHeader({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        "Catalogue App".text.xl4.bold.color(MyTheme.darkBluishColor).make(),
+        "Trending Products".text.xl2.make(),
+      ],
+    );
+  }
+}
