+++ Flutter/flutter_catalogue/lib/pages/home_widgets/catalogue_image.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_catalogue/widgets/themes.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+class CatalogueImage extends StatelessWidget {
+  final String image;
+
+  const CatalogueImage({Key? key, required this.image}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Image.network(
+      image,
+    ).box.rounded.p8.color(MyTheme.creamColor).make().p16().w40(context);
+  }
+}
