+++ Flutter/flutter_catalogue/lib/pages/home_widgets/catalogue_list.dart
@@ -1,0 +1,65 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_catalogue/models/catalogue.dart';
+import 'package:flutter_catalogue/pages/home_widgets/catalogue_image.dart';
+import 'package:flutter_catalogue/widgets/themes.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+class CatalogueList extends StatelessWidget {
+  const CatalogueList({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ListView.builder(
+      shrinkWrap: true,
+      itemCount: CatalogueModel.items.length,
+      itemBuilder: (context, index) {
+        final catalogue = CatalogueModel.items[index];
+        return CatalogueItem(catalogue: catalogue);
+      },
+    );
+  }
+}
+
+class CatalogueItem extends StatelessWidget {
+  final Item catalogue;
+
+  const CatalogueItem({Key? key, required this.catalogue}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return VxBox(
+      child: Row(
+        children: [
+          CatalogueImage(
+            image: catalogue.image,
+          ),
+          Expanded(
+              child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              catalogue.name.text.lg.color(MyTheme.darkBluishColor).bold.make(),
+              catalogue.desc.text /*textStyle(context.captionStyle)*/ .make(),
+              10.heightBox,
+              ButtonBar(
+                alignment: MainAxisAlignment.spaceBetween,
+                buttonPadding: EdgeInsets.zero,
+                children: [
+                  "\$${catalogue.price}".text.bold.xl.make(),
+                  ElevatedButton(
+                    onPressed: () {},
+                    style: ButtonStyle(
+                        backgroundColor:
+                            MaterialStateProperty.all(MyTheme.darkBluishColor),
+                        shape: MaterialStateProperty.all(StadiumBorder())),
+                    child: "Buy".text.make(),
+                  )
+                ],
+              ).pOnly(right: 8.0)
+            ],
+          ))
+        ],
+      ),
+    ).white.roundedLg.square(150).make().py16();
+  }
+}
